Changed Person and User classes to interfaces
