+++ src/main/java/com/unirutas/models/Driver.java
@@ -1,11 +1,33 @@
 package com.unirutas.models;
 
-import java.util.List;
+import com.unirutas.annotations.Column;
+import com.unirutas.annotations.PrimaryKey;
+import com.unirutas.annotations.Table;
 
-public class Driver extends Person {
-    private List<Bus> drivers;
+@Table(name = "Driver")
+@PrimaryKey({"document_number"})
+public class Driver implements Person {
+    @Column(name = "name")
+    private String name;
+    private String documentNumber;
 
-    public Driver(String name, String document_number) {
-        super(name, document_number);
+    public Driver(String name, String documentNumber) {
+        this.name = name;
+        this.documentNumber = documentNumber;
+    }
+
+    @Override
+    public String getName() {
+        return name;
+    }
+
+    @Override
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    @Override
+    public String getCode() {
+        return documentNumber;
     }
 }
