+++ src/main/java/com/unirutas/models/Person.java
@@ -1,29 +1,11 @@
 package com.unirutas.models;
 
-public class Person {
-    private String name;
-    private String code;
+public interface Person {
+    String getName();
 
-    public Person(String name, String code) {
-        this.name = name;
-        this.code = code;
-    }
+    public void setName(String name);
 
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getCode() {
-        return code;
-    }
-
-    public void setCode(String code) {
-        this.code = code;
-    }
+    public String getCode();
 
 }
