+++ src/main/java/com/unirutas/models/User.java
@@ -1,29 +1,13 @@
 package com.unirutas.models;
 
-public abstract class User extends Person {
-    private final String username;
-    private String password;
+import com.unirutas.annotations.Column;
 
-    public User(String name, String code, String username, String password) {
-        super(name, code);
-        this.username = username;
-        this.password = password;
-    }
+public interface User extends Person {
 
-    public String getUsername() {
-        return username;
-    }
+    public String getUsername();
 
-    public String getPassword() {
-        return password;
-    }
+    public String getPassword();
 
-    public void changePassword(String newPassword) {
-        this.password = newPassword;
-    }
-
-    public void insert(User user) {}
-
-    public void update(User user) {}
+    public void changePassword(String newPassword);
 
 }
